+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Python class for reading and writing
 
 ## unreleased
 
+- `NLLGrid.recenter()` method to move origin of grid cartesian system to the
+  grid center
 - Added `TRANS_MERC` and `AZIMUTHAL_EQUIDIST` projections
 - Fix projections when `map_rot != 0`